move init of store inside the app.vue
